Update script for ROC-curve and combined results of 3 networks
- Update existing validation script to plot ROC-curve for individual
molecules and for combined all molecules
- New script for specifically generate the results from combined
modelling results of the 3 networks using the rule of majority (without
zero) by combining rowSums and take the sign of sum results

Note1: Might still need to check carefully why the results seem to be
too good to be true for combined ones with thresholds.

Note2: Results are saved in separated folders to save committing time.