Update script for ROC-curve and combined results of 3 networks
- Update existing validation script to plot ROC-curve for individual
molecules and for combined all molecules
- New script for specifically generate the results from combined
modelling results of the 3 networks using the rule of majority (without
zero) by combining rowSums and take the sign of sum results

Note1: Might still need to check carefully why the results seem to be
too good to be true for combined ones with thresholds.

Note2: Results are saved in separated folders to save committing time.

Former-commit-id: 1ff3b4220a1ccc5911c103699dcc688a216bab91 [formerly 156eaf47c95a18ab78f03a42ce191ff198b9716b] [formerly dbca7361aa3006485213a8e9407b8e6474fa6030 [formerly 0981e9514beeacd30abdcc55be2079d6f9f4e824]]
Former-commit-id: c35a2b1979e9a38808e9b6e88318114666d387a3
Former-commit-id: 68ff55e5dabd52f541350cc21331ad653244cc8a