Update script for ROC-curve and combined results of 3 networks
- Update existing validation script to plot ROC-curve for individual
molecules and for combined all molecules
- New script for specifically generate the results from combined
modelling results of the 3 networks using the rule of majority (without
zero) by combining rowSums and take the sign of sum results

Note1: Might still need to check carefully why the results seem to be
too good to be true for combined ones with thresholds.

Note2: Results are saved in separated folders to save committing time.

Former-commit-id: dbca7361aa3006485213a8e9407b8e6474fa6030 [formerly 0981e9514beeacd30abdcc55be2079d6f9f4e824]
Former-commit-id: 156eaf47c95a18ab78f03a42ce191ff198b9716b