Fix issue with measurementsWeight mapping
- The indexing of absDiff starts from measurements first followed by
non-measured nodes e.g. 8 9 10 1 2 3 4 5 6 7 for Ex5. Therefore, the
mapping via expNodesReduced also have to change according to the
position of absDiff (stored in “idxMeasured”).

- Add measurementsWeights in Ex3 as a case study and obtained correct
results [!!NOTE!!] CHANGE OF THE CONCEPT - The measurements with higher
(DoRothEA) score -> have MORE weights (instead of less weights) ->
Therefore cplex try to fit the one with higher weights first to lower
the overall objective function — Also, the weights are now considered
as an “add-on” part instead of “soften” part in the objective function
so that it will not interfere with the influence from beta once the the
ratio alpha/beta is low.


Former-commit-id: dcbc82d7217bd5a1fa89a730566891616b52934e [formerly 44483b31cf818961bff8b71d1560d579d4879dee] [formerly 04a319fac660e45feee9a3e930cfae239fdcbb10 [formerly c0b260d66be2296e968e77068f08dd728384b3aa]]
Former-commit-id: 6ec4122a9c7145807ce937ce692859ae567d796a
Former-commit-id: 57b89048708bb3c9dbb261650de168cda9673805