Revise CARNIVAL input format and clean up Uniprot2GeneSymbol
- According to the received feedbacks, specifying the input variables
directly in R would be easier and less error-prone for the users
comparing to the modifying a text file. Also, it would be easier for
the parallelisation on the cluster.

- Now the “runCARNIVAL” function takes all input arguments directly.
Descriptions for all variables are available in the help file. The QC
step for input format is also introduced to ensure that the user
plugged in the correct set of input information.

- The vignette of CARNIVAL is updated accordingly. An example to run
CARNIVAL with built-in in example and user-provided set of inputs are
available.

- Mod: clean up to Uniprot2GeneSymbol.R script to summarise and export
a unique set of unmapped node names to gene symbol instead of reporting
them on multiple lines.


Former-commit-id: 4457a0f3af449742152d4599785f0b6edefc5b11 [formerly b94e0cb7d87b6d48fff36255f25e96278f501322] [formerly e3bcb40c7f8086308970b121e9e46a3eebfdd328 [formerly a7793064f744c1e16482cb6d31e575344267c1a7]]
Former-commit-id: a4fc2dbcd91d6c8df62f067de32bf9a4287e716b
Former-commit-id: 559ace368ffc44b5448076f94dfafc533cef9914