Update runCARNIVAL.R
Issue: Upon running CARNIVAL on the cluster with a separated (driver) R-script, users couldn't directly pass the arguments from the cluster (argsJob) directly to the runCARNIVAL function.

Fix: Replaced the parameter "parallelCR" (T/F) with two new parameters "parallelIdx1" and "parallelIdx2" to allow a direct assignment of the indexing-number of intermediate and result files which should be unique on the cluster. Note that the user have to include the following lines on their driver script to allow parallelisation

library(doParallel)
argsJob=commandArgs(trailingOnly = TRUE)

Also, the runCARNIVAL function needs additional assignment for the new parameters:

runCARNIVAL(parallelIdx1 = as.numeric(argsJob[1]),
                              parallelIdx2 = as.numeric(argsJob[2]))

This new implementation will also be documented in the README and Wiki section upon merging.


Former-commit-id: 627542df63a8c23cb6a7627d7cf323ad8540933a [formerly bd0feae7e0b910e2b0a03eca9f55caf236470ae8] [formerly 99d4a1ea77fbcdd5801355a1bb1b619612adbdbc [formerly f91ac4c4ab31a24c7edc81f0c2349b004b7a4228]]
Former-commit-id: 4c6c6d0c7c9e0e9df695698de4fa65608b89030a
Former-commit-id: a85fb8fef5c7fa0355b2e876979688a3cfa5217e