Update runDoRothEA.R
Fix the hidden bug when calling the object "viper_regulon" instead of "regulon" inside the function.

This function didn't return an error so far, possible due to the fact that the object "viper_regulon" was loaded into the workspace prior to calling the function. Nevertheless, it is a standard practice to correct all detected bugs to prevent future issues.


Former-commit-id: 84ef4f698328f54db0e49e1f83e2cb1a98292778 [formerly 68fc0c6725364f92026af7460af2bd7f16e8e58d]
Former-commit-id: ef1cc7d04be8abfaf0cc843e5c4b4f376ff83376