Update runDoRothEA.R
Fix the hidden bug when calling the object "viper_regulon" instead of "regulon" inside the function.

This function didn't return an error so far, possible due to the fact that the object "viper_regulon" was loaded into the workspace prior to calling the function. Nevertheless, it is a standard practice to correct all detected bugs to prevent future issues.