Update runDoRothEA.R
Fix the hidden bug when calling the object "viper_regulon" instead of "regulon" inside the function.

This function didn't return an error so far, possible due to the fact that the object "viper_regulon" was loaded into the workspace prior to calling the function. Nevertheless, it is a standard practice to correct all detected bugs to prevent future issues.


Former-commit-id: 31556c2ee8087353495ff013625035f07b61b221 [formerly ef1cc7d04be8abfaf0cc843e5c4b4f376ff83376] [formerly 84ef4f698328f54db0e49e1f83e2cb1a98292778 [formerly 68fc0c6725364f92026af7460af2bd7f16e8e58d]]
Former-commit-id: a8070e06665f867998bc580f95d4061ef8753ef7
Former-commit-id: a42bf2bc4ba28f91cca127e876b0055bc5e28b76